# apps/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "文件管理系统后端 API，支持大文件分片上传、下载和预览",
        "title": "文件管理系统 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@filemanager.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/documents": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "分页获取文档列表，支持关键词搜索",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文档管理"
                ],
                "summary": "获取文档列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "搜索关键词",
                        "name": "keyword",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/service.ListDocumentsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/documents/chunks/complete": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "合并所有分片并完成上传",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文档上传"
                ],
                "summary": "完成上传",
                "parameters": [
                    {
                        "description": "完成上传信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.CompleteUploadRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/service.CompleteUploadResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/documents/chunks/init": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "初始化一个大文件分片上传",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文档上传"
                ],
                "summary": "初始化分片上传",
                "parameters": [
                    {
                        "description": "上传信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.InitUploadRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/service.InitUploadResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/documents/chunks/upload": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "上传单个分片",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文档上传"
                ],
                "summary": "上传分片",
                "parameters": [
                    {
                        "type": "string",
                        "description": "上传ID",
                        "name": "upload_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "分片序号",
                        "name": "chunk_number",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "分片文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/documents/chunks/{upload_id}": {
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "取消分片上传并清理临时文件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文档上传"
                ],
                "summary": "取消上传",
                "parameters": [
                    {
                        "type": "string",
                        "description": "上传ID",
                        "name": "upload_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/documents/upload": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "直接上传单个文件（适用于小文件）",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文档上传"
                ],
                "summary": "单文件上传",
                "parameters": [
                    {
                        "type": "file",
                        "description": "文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/service.UploadSingleFileResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误或文件类型不支持",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "413": {
                        "description": "文件大小超过限制",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/documents/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "根据ID获取文档详细信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文档管理"
                ],
                "summary": "获取文档信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文档ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Document"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "删除指定文档",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文档管理"
                ],
                "summary": "删除文档",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文档ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/documents/{id}/download": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "下载指定文档",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "文档访问"
                ],
                "summary": "下载文档",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文档ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
        "/documents/{id}/preview": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "在线预览文档（支持图片和PDF）",
                "produces": [
                    "image/jpeg",
                    "application/pdf"
                ],
                "tags": [
                    "文档访问"
                ],
                "summary": "预览文档",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文档ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "model.Document": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "file_extension": {
                    "type": "string"
                },
                "file_name": {
                    "type": "string"
                },
                "file_size": {
                    "type": "integer"
                },
                "file_type": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "md5_hash": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "storage_path": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "upload_id": {
                    "type": "string"
                }
            }
        },
...
        "service.CompleteUploadRequest": {
            "type": "object",
            "required": [
                "account",
                "upload_id"
            ],
            "properties": {
                "account": {
                    "description": "应用账号或用户名",
                    "type": "string"
                },
                "upload_id": {
                    "type": "string"
                }
            }
        },
        "service.CompleteUploadResponse": {
            "type": "object",
            "properties": {
                "document_id": {
                    "type": "string"
                },
                "file_name": {
                    "type": "string"
                },
                "file_size": {
                    "type": "integer"
                }
            }
        },
...
        "service.InitUploadRequest": {
            "type": "object",
            "required": [
                "chunk_size",
                "file_name",
                "file_size"
            ],
            "properties": {
                "chunk_size": {
                    "type": "integer"
                },
                "file_name": {
                    "type": "string"
                },
                "file_size": {
                    "type": "integer"
                }
            }
        },
        "service.InitUploadResponse": {
            "type": "object",
            "properties": {
                "chunk_size": {
                    "type": "integer"
                },
                "file_size": {
                    "type": "integer"
                },
                "total_chunks": {
                    "type": "integer"
                },
                "upload_id": {
                    "type": "string"
                }
            }
        },
...
        "service.ListDocumentsResponse": {
            "type": "object",
            "properties": {
                "documents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Document"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "service.UploadSingleFileResponse": {
            "type": "object",
            "properties": {
                "document_id": {
                    "type": "string"
                },
                "file_name": {
                    "type": "string"
                },
                "file_size": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "请输入 Bearer {token} 格式的 Token，例如：Bearer eyJhbGc...",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}